+++ backend/app/api/routes/authentication.py
@@ -1,6 +1,10 @@
-from fastapi import APIRouter, Body, Depends, HTTPException
-from starlette.status import HTTP_201_CREATED, HTTP_400_BAD_REQUEST
+from typing import List
 
+from fastapi import APIRouter, Body, Depends, HTTPException
+from starlette.status import HTTP_201_CREATED, HTTP_400_BAD_REQUEST, HTTP_403_FORBIDDEN
+
+from app.api.dependencies.authentication import get_current_user_authorizer
+from app.models.domain.users import User
 from app.api.dependencies.database import get_repository
 from app.core.config import get_app_settings
 from app.core.settings.app import AppSettings
@@ -12,13 +16,28 @@ from app.models.schemas.users import (
     UserInResponse,
     UserWithToken,
 )
+from app.models.domain.users import User, UserRole
+
 from app.resources import strings
 from app.services import jwt
 from app.services.authentication import check_email_is_taken, check_username_is_taken
 from app.services.event import send_event
 
+
 router = APIRouter()
 
+@router.get("/", response_model=List[User], name="users:get-all-users")
+async def retrieve_all_users(
+    user: User = Depends(get_current_user_authorizer()),
+    users_repo: UsersRepository = Depends(get_repository(UsersRepository)),
+) -> List[User]:
+    if user.role != UserRole.admin:
+        raise HTTPException(                                                                                        
+                status_code=HTTP_403_FORBIDDEN,
+                detail=strings.UNAUTHORIZED,
+            )
+    users = await users_repo.get_all_users()
+    return users
 
 @router.post("/login", response_model=UserInResponse, name="auth:login")
 async def login(
